requirements.txt
numpy==1.21.2
Pillow==8.3.2
tweepy==4.1.0
selenium==3.141.0
webdriver-manager==3.5.2